Updated README files for all Assignements from 1 to 6 and also included some minor code related changes

+++ ASSIGNMENT 1/Task 1.py
@@ -1,5 +1,5 @@
-a = int(input("Enter the first number: "))
-b = int(input("Enter the second number: "))
+a = float(input("Enter the first number: "))
+b = float(input("Enter the second number: "))
 
 print("Addition:", a + b)
 print("Subtraction:", a - b)
